+++ docs/CHANGELOG.md
@@ -1,7 +1,8 @@
 # Version 2.4.1 2018-04-11
 - Fix `\CfdiUtils\Certificado\Certificado` when reading serial number.
-  - Use `serialNumberHex` if available, if not then use `serialNumber` and convert to hex using bcdechex
-  - Add test to this condition. 
+  - Use `serialNumberHex` if available, if not then use `serialNumber` and convert to hex.
+- Move serial number string conversion to class `\CfdiUtils\Certificado\SerialNumber`.
+  This class is not for public use but for use inside `Certificate`.
 
 # Version 2.4.0 2018-02-08
 - Add the feature to order the children nodes for a `CfdiUtils\Nodes\Nodes` object.